+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project closely adheres to [Semantic Versioning](https://semver.org/spe
 
 ## 1.2.2
 - Chore - make lint also run ruff format --check
+- Add - tags command to list tags and affected scenario counts
 
 ## 1.2.1
 - Add - tab information to html report and cucu debug console log